Move cuda_stream_view.hpp out of detail folder
