+++ include/rmm/mr/device/device_memory_resource.hpp
@@ -15,8 +15,8 @@
  */
 #pragma once
 
+#include <rmm/cuda_stream_view.hpp>
 #include <rmm/detail/aligned.hpp>
-#include <rmm/detail/cuda_stream_view.hpp>
 
 #include <cstddef>
 #include <utility>
