+++ tests/cuda_stream_tests.cu
@@ -16,7 +16,7 @@
 
 #include "gtest/gtest.h"
 
-#include <rmm/detail/cuda_stream_view.hpp>
+#include <rmm/cuda_stream_view.hpp>
 #include <rmm/device_buffer.hpp>
 
 #include <cuda_runtime_api.h>
